refactor: eliminate duplication with BaseChromiumCookieQueryStrategy
- Create BaseChromiumCookieQueryStrategy with all shared Chromium logic
- Chrome and Chromium strategies now only implement path discovery
- Reduced code duplication from ~300 lines to ~60 lines per strategy
- Shared logic includes:
  - Platform support checking
  - Cookie file processing with retry
  - Browser lock handling
  - Cookie decryption and error handling
  - Meta version retrieval from database
- Class inheritance clearly models browser relationships
- Fix TypeScript type errors in BrowserLockHandler

This massive DRY refactoring eliminates ~80% code duplication between
Chrome and Chromium strategies while maintaining full functionality and
improving maintainability through proper abstraction.
